requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.7.0
APScheduler==3.6.3
async-timeout==5.0.1
attrs==24.3.0
cachetools==4.2.2
certifi==2024.12.14
click==8.1.8
deprecation==2.1.0
et_xmlfile==2.0.0
exceptiongroup==1.2.2
fastapi==0.115.6
frozenlist==1.5.0
gotrue==2.11.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
multidict==6.1.0
nest-asyncio==1.6.0
numpy==2.0.2
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
postgrest==0.18.0
propcache==0.2.1
pydantic==2.10.4
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.9
pytz==2024.2
realtime==2.0.6
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
storage3==0.9.0
StrEnum==0.4.15
supabase==2.10.0
supafunc==0.7.0
tornado==6.4.2
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
uvicorn==0.34.0
websockets==13.1
yarl==1.18.3
